+++ src/comments/repositories/comments.repository.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateCommentDto } from '../dto/create-comment.dto';
 import { CommentEntity } from '../entities/comment.entity';
@@ -8,7 +12,11 @@ import { UpdateCommentDto } from '../dto/update-comment.dto';
 export class CommentsRepository {
   constructor(private readonly prisma: PrismaService) {}
 
-  async create(createCommentDto: CreateCommentDto): Promise<CommentEntity> {
+  async create(
+    createCommentDto: CreateCommentDto,
+    token_id: string,
+  ): Promise<CommentEntity> {
+    createCommentDto = { ...createCommentDto, user_id: token_id };
     return this.prisma.comment.create({
       data: createCommentDto,
     });
@@ -30,9 +38,23 @@ export class CommentsRepository {
     return findComment;
   }
 
+  async findByAnnouncement(id: string): Promise<CommentEntity[]> {
+    const findComments = await this.prisma.comment.findMany({
+      where: {
+        announcement_id: id,
+      },
+    });
+    if (!findComments) {
+      throw new NotFoundException('Comment not found');
+    }
+    return findComments;
+  }
+
   async update(
     id: string,
     updateCommentDto: UpdateCommentDto,
+    token_id: string,
+    type: string,
   ): Promise<CommentEntity> {
     const findComment = await this.prisma.comment.findUnique({
       where: {
@@ -42,15 +64,27 @@ export class CommentsRepository {
     if (!findComment) {
       throw new NotFoundException('Comment not found');
     }
-    return this.prisma.comment.update({
-      where: {
-        id,
-      },
-      data: updateCommentDto,
-    });
+    if (type == 'COMPRADOR') {
+      if (findComment.user_id == token_id) {
+        return this.prisma.comment.update({
+          where: {
+            id,
+          },
+          data: updateCommentDto,
+        });
+      } else {
+        throw new UnauthorizedException(
+          'Only owner can perform this operation',
+        );
+      }
+    }
   }
 
-  async remove(id: string): Promise<CommentEntity> {
+  async remove(
+    id: string,
+    token_id: string,
+    type: string,
+  ): Promise<CommentEntity> {
     const findComment = await this.prisma.comment.findUnique({
       where: {
         id,
@@ -59,10 +93,34 @@ export class CommentsRepository {
     if (!findComment) {
       throw new NotFoundException('Comment not found');
     }
-    return this.prisma.comment.delete({
+    if (type == 'COMPRADOR') {
+      if (findComment.user_id == token_id) {
+        return this.prisma.comment.delete({
+          where: {
+            id,
+          },
+        });
+      } else {
+        throw new UnauthorizedException(
+          'Only the owner of the ad or comment can perform this operation',
+        );
+      }
+    }
+    const verifyOwner = await this.prisma.announcement.findUnique({
       where: {
-        id,
+        id: findComment.announcement_id,
       },
     });
+    if (verifyOwner.user_id == token_id) {
+      return this.prisma.comment.delete({
+        where: {
+          id,
+        },
+      });
+    } else {
+      throw new UnauthorizedException(
+        'Only the owner of the ad or comment can perform this operation',
+      );
+    }
   }
 }
